Type ahead search for filters (#23)
For the location filters, we want to be able to have type-ahead
searching to easily find a country/state/city.

In order to do this, I switched to using bootstrap-select &
selectpicker. I updated all filters to use the selectpicker so that they
look the same visually, even if they don't use type-ahead search.